requirements.txt
aws-cdk-lib==2.78.0
constructs>=10.0.0,<11.0.0
pandas==2.0.1
tqdm==4.65.0
